refactor(event_dag): move layer machinery from accumulator.rs to layers.rs
Pure mechanical move, no behavior change. EventLayers, EventLayer, and
their tests join CausalRelation in layers.rs, which previously held only
that enum while the layer types squatted in accumulator.rs.
accumulator.rs keeps the comparison-phase pieces: EventAccumulator,
ComparisonResult, and the shared DAG-walk helpers.

Both types now re-export from the event_dag module root so consumers
import crate::event_dag::{EventLayer, CausalRelation} instead of
encoding the internal file layout; the backward CausalRelation re-export
through accumulator is removed. EventLayers reads the accumulated DAG
through the existing dag() accessor rather than the private field it
could reach when co-located.
