+++ core/src/event_dag/mod.rs
@@ -14,6 +14,7 @@ pub(crate) mod relation;
 mod tests;
 
 pub(crate) use comparison::compare;
+pub(crate) use layers::{CausalRelation, EventLayer};
 pub(crate) use relation::AbstractCausalRelation;
 
 /// Default budget for DAG traversal — large enough for typical histories
