+++ core/src/property/backend/mod.rs
@@ -8,7 +8,7 @@ pub mod lww;
 //pub mod pn_counter;
 pub mod yrs;
 use crate::error::{MutationError, RetrievalError, StateError};
-use crate::event_dag::accumulator::EventLayer;
+use crate::event_dag::EventLayer;
 pub use lww::LWWBackend;
 //pub use pn_counter::PNBackend;
 pub use yrs::YrsBackend;
